Revert "Class changes"
